pyproject.toml
[tool.poetry]
name = "object-detection"
version = "0.1.0"
description = ""
authors = ["hfmartinez <heberthfabianmartinez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.1.1"
sqlalchemy = "^2.0.28"
opencv-python = "^4.9.0.80"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
requests = "^2.31.0"
fastapi-sqlalchemy = "^0.2.1"
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.110.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
